% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcr.R
\name{fcr}
\alias{fcr}
\title{Fuzzy Censored Regression}
\usage{
fcr(Y,X,lower=c(L1,L2,L3),upper=c(U1,U2,U3),type)
}
\arguments{
\item{y}{is an output fuzzy number}

\item{x}{is an input fuzzy number}

\item{lower}{is a set contains lower end censored observations of left, centre and right of triangular fuzzy number.Default lower limit is zero "0".}

\item{upper}{is a set contains Upper end censored observations of left, centre and right of triangular fuzzy number.Default upper limit is infinity "Inf".}

\item{type}{is model to be used (1 or 2)}

\item{1}{"Fuzzy output, Fuzzy input and Fuzzy Parameters"}

\item{2}{"Fuzzy output, Crisp input and Fuzzy Parameters"}
}
\description{
It gives the estimates of Fuzzy Censored Regression by using the iterative based method of Newton Raphson.
}
\examples{
If given Triangular Fuzzy NUmber

library("VGAM")

x_left<-c(1.5,3.0,4.5,6.5,8.0,9.5,10.5,12.0)
x_centre<-c(2.0,3.5,5.5,7.0,8.5,10.5,11.0,12.5)
x_right<-c(2.5,4.0,6.5,7.5,9.0,11.5,11.5,13.0)

y_left<-c(3.5,5.0,6.5,6.0,8.0,7.0,10.0,9.0)
y_centre<-c(4.0,5.5,7.5,6.5,8.5,8.0,10.5,9.5)
y_right<-c(4.5,6.0,8.5,7.0,9.0,9.0,11.0,10.0)

X<-cbind(x_left,x_centre,x_right)
Y<-cbind(y_left,y_centre,y_right)

fcr(Y,X,lower=c(2.0,3.5,5.0),upper=c(8.5,9.0,10.0),type=1)


}
\author{
Mohsin Shahzad
}
