% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flr.R
\name{flr}
\alias{flr}
\title{Fuzzy Linear Regression}
\usage{
flr(Y,X,type)
}
\arguments{
\item{type}{is model to be used (1 or 2)}

\item{x}{is an input fuzzy number}

\item{y}{is an output fuzzy number}

\item{1}{"Fuzzy output, Fuzzy input and Fuzzy Parameters"}

\item{2}{"Fuzzy output, Crisp input and Fuzzy Parameters"}
}
\description{
It gives the estimates of fuzzy regression by using the method of least squares.
}
\examples{
If given Triangular Fuzzy NUmber

x_left<-c(1.5,3.0,4.5,6.5,8.0,9.5,10.5,12.0)
x_centre<-c(2.0,3.5,5.5,7.0,8.5,10.5,11.0,12.5)
x_right<-c(2.5,4.0,6.5,7.5,9.0,11.5,11.5,13.0)

y_left<-c(3.5,5.0,6.5,6.0,8.0,7.0,10.0,9.0)
y_centre<-c(4.0,5.5,7.5,6.5,8.5,8.0,10.5,9.5)
y_right<-c(4.5,6.0,8.5,7.0,9.0,9.0,11.0,10.0)

X<-cbind(x_left,x_centre,x_right)
Y<-cbind(y_left,y_centre,y_right)

flr(Y,X,type=1)


}
\author{
Mohsin Shahzad
}
