% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqr.R
\name{fqr}
\alias{fqr}
\title{Fuzzy Quantile Regression}
\usage{
fqr(X,y_left,y_centre,y_right,t,type)
}
\arguments{
\item{t}{is a spesified quantile ranges from 0 to 1 i.e t=[0,1]}

\item{type}{spesifies the model (1 or 2)}

\item{X}{is an input fuzzy number}

\item{y_left}{is an left output fuzzy number}

\item{y_centre}{is an centre of output fuzzy number}

\item{y_right}{is an right output fuzzy number}

\item{1}{"Fuzzy output, Fuzzy input and Fuzzy Parameters"}

\item{2}{"Fuzzy output, Crisp input and Fuzzy Parameters"}
}
\description{
It gives the estimates of fuzzy quantile regression using the method of Weighted Least Absolute Deviation (WLAD). It converts the input variables into Linear Programming Problem (LPP) and uses the Simplex Algorithm to solve the LPP.
}
\examples{
If given Triangular Fuzzy NUmber

library ("lpSolve")

x_left<-c(1.5,3.0,4.5,6.5,8.0,9.5,10.5,12.0)
x_centre<-c(2.0,3.5,5.5,7.0,8.5,10.5,11.0,12.5)
x_right<-c(2.5,4.0,6.5,7.5,9.0,11.5,11.5,13.0)

y_left<-c(3.5,5.0,6.5,6.0,8.0,7.0,10.0,9.0)
y_centre<-c(4.0,5.5,7.5,6.5,8.5,8.0,10.5,9.5)
y_right<-c(4.5,6.0,8.5,7.0,9.0,9.0,11.0,10.0)

X<-cbind(x_left,x_centre,x_right)

t<-0.5
fqr(X,y_left,y_centre,y_right,t,type=1)

}
\author{
Mohsin Shahzad
}
